requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
Django==4.0.2
docopt==0.6.2
Flask==2.0.3
gunicorn==20.1.0
idna==3.3
imageio==2.16.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
networkx==2.6.3
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.1
pipreqs==0.4.11
pyparsing==3.0.7
pytesseract==0.3.8
PyWavelets==1.2.0
requests==2.27.1
scikit-image==0.19.1
scipy==1.8.0
sqlparse==0.4.2
tifffile==2022.2.9
tzdata==2021.5
urllib3==1.26.8
Werkzeug==2.0.3
yarg==0.1.9
